chore: rename rule to zeebe-user-task
